pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-log-factory-ext-otel"
version = "1.5.0rc1"
description = "OpenTelemetry log handler and tracing plugin for simple_log_factory"
readme = "README.md"
license = "GPL-3.0-only"
license-files = ["LICENSE.md"]
requires-python = ">=3.9"
authors = [
    {name = "Breno RdV"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Logging",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["logging", "opentelemetry", "otel", "observability", "simple-log-factory", "tracing", "traces", "spans"]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.optional-dependencies]
psycopg2 = ["opentelemetry-instrumentation-psycopg2>=0.60b1"]
psycopg = ["opentelemetry-instrumentation-psycopg>=0.60b1"]
db = [
    "opentelemetry-instrumentation-psycopg2>=0.60b1",
    "opentelemetry-instrumentation-psycopg>=0.60b1",
]
requests = ["opentelemetry-instrumentation-requests>=0.60b1"]
fastapi = ["opentelemetry-instrumentation-fastapi>=0.60b1"]
cross-resource = [
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.15.1",
    "simple-log-factory>=1.0.0",
    "opentelemetry-instrumentation-psycopg2>=0.60b1",
    "opentelemetry-instrumentation-psycopg>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
]

[project.urls]
Homepage = "https://github.com/raccoon/simple-log-factory-ext-otel"
Repository = "https://github.com/raccoon/simple-log-factory-ext-otel"
Issues = "https://github.com/raccoon/simple-log-factory-ext-otel/issues"

[tool.setuptools.packages.find]
include = ["simple_log_factory_ext_otel*"]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
strict = true

[[tool.mypy.overrides]]
module = ["opentelemetry.instrumentation.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
asyncio_mode = "auto"
